# 2.AssetsAI/1.RefineryDigitalTwin/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Refinery Digital Twin — REST
  description: Read-only HTTP access to live + historical DWSIM snapshots from the
    Stage 2 streamer. No DWSIM connection, no cloud — just the JSONL stream the streamer
    writes to local disk.
  version: 0.1.0
paths:
...
  /snapshots/range:
    get:
      tags:
      - snapshots
      summary: Snapshots Range
      operationId: snapshots_range_snapshots_range_get
      parameters:
      - name: since
        in: query
        required: true
        schema:
          type: string
          description: ISO 8601 lower bound (inclusive)
          title: Since
        description: ISO 8601 lower bound (inclusive)
      - name: until
        in: query
        required: true
        schema:
          type: string
          description: ISO 8601 upper bound (inclusive)
          title: Until
        description: ISO 8601 upper bound (inclusive)
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 10000
          minimum: 1
          default: 1000
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RangeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /tags/{tag_id}:
    get:
      tags:
      - tags
      summary: Tags Get
      operationId: tags_get_tags__tag_id__get
      parameters:
      - name: tag_id
        in: path
        required: true
        schema:
          type: string
          title: Tag Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagEntry'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tags/{tag_id}/value:
    get:
      tags:
      - tags
      summary: Tags Value
      operationId: tags_value_tags__tag_id__value_get
      parameters:
      - name: tag_id
        in: path
        required: true
        schema:
          type: string
          title: Tag Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagValueResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tags/{tag_id}/history:
    get:
      tags:
      - tags
      summary: Tags History
      operationId: tags_history_tags__tag_id__history_get
      parameters:
      - name: tag_id
        in: path
        required: true
        schema:
          type: string
          title: Tag Id
      - name: since
        in: query
        required: true
        schema:
          type: string
          description: ISO 8601 lower bound (inclusive)
          title: Since
        description: ISO 8601 lower bound (inclusive)
      - name: until
        in: query
        required: true
        schema:
          type: string
          description: ISO 8601 upper bound (inclusive)
          title: Until
        description: ISO 8601 upper bound (inclusive)
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 10000
          minimum: 1
          default: 1000
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagHistoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /ontology/entities/{entity_id}:
    get:
      tags:
      - ontology
      summary: Ontology Entity
      description: 'One entity + its outbound `relationships`, the reverse-direction

        `inbound_relationships`, and the `tag_ids` mapped to it from

        tag_mapping.json. 404 if unknown.'
      operationId: ontology_entity_ontology_entities__entity_id__get
      parameters:
      - name: entity_id
        in: path
        required: true
        schema:
          type: string
          title: Entity Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OntologyEntityDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ontology/tags/{tag_id}:
    get:
      tags:
      - ontology
      summary: Ontology Tag
      description: 'Reverse lookup: which entity owns this tag? Augments the tag

        mapping with the owning entity''s name and type.'
      operationId: ontology_tag_ontology_tags__tag_id__get
      parameters:
      - name: tag_id
        in: path
        required: true
        schema:
          type: string
          title: Tag Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OntologyTagInfo'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ontology/resolve:
    get:
      tags:
      - ontology
      summary: Ontology Resolve
      description: 'Natural-language term → ranked list of {entity, tag_ids}. Demo:

        `?term=condenser%20duty` → ES-CONDENSER_DUTY with EnergyFlow tag.'
      operationId: ontology_resolve_ontology_resolve_get
      parameters:
      - name: term
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          description: 'Natural-language term to resolve. Two-tier match: exact alias
            hit (score 100) → substring fallback (score scaled by length ratio, capped
            95).'
          title: Term
        description: 'Natural-language term to resolve. Two-tier match: exact alias
          hit (score 100) → substring fallback (score scaled by length ratio, capped
          95).'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OntologyResolveResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OntologyEntityDetail:
      properties:
        id:
          type: string
          title: Id
        type:
          type: string
          title: Type
        name:
          type: string
          title: Name
        aliases:
          items:
            type: string
          type: array
          title: Aliases
        subsystem:
          anyOf:
          - type: string
          - type: 'null'
          title: Subsystem
        properties:
          additionalProperties: true
          type: object
          title: Properties
        stages:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Stages
        normal_ranges:
          anyOf:
          - additionalProperties:
              additionalProperties: true
              type: object
            type: object
          - type: 'null'
          title: Normal Ranges
        relationships:
          items:
            $ref: '#/components/schemas/OntologyRelationship'
          type: array
          title: Relationships
        inbound_relationships:
          items:
            $ref: '#/components/schemas/OntologyInboundRelationship'
          type: array
          title: Inbound Relationships
        tag_ids:
          items:
            type: string
          type: array
          title: Tag Ids
      type: object
      required:
      - id
      - type
      - name
      title: OntologyEntityDetail
    OntologyInboundRelationship:
      properties:
        type:
          type: string
          title: Type
        source:
          type: string
          title: Source
      type: object
      required:
      - type
      - source
      title: OntologyInboundRelationship
    OntologyRelationship:
      properties:
        type:
          type: string
          title: Type
        target:
          type: string
          title: Target
      type: object
      required:
      - type
      - target
      title: OntologyRelationship
    OntologyResolveHit:
      properties:
        entity_id:
          type: string
          title: Entity Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
        match_type:
          type: string
          title: Match Type
        matched_phrase:
          type: string
          title: Matched Phrase
        score:
          type: integer
          title: Score
        tag_ids:
          items:
            type: string
          type: array
          title: Tag Ids
      type: object
      required:
      - entity_id
      - match_type
      - matched_phrase
      - score
      - tag_ids
      title: OntologyResolveHit
    OntologyResolveResponse:
      properties:
        term:
          type: string
          title: Term
        count:
          type: integer
          title: Count
        results:
          items:
            $ref: '#/components/schemas/OntologyResolveHit'
          type: array
          title: Results
      type: object
      required:
      - term
      - count
      - results
      title: OntologyResolveResponse
    OntologyTagInfo:
      properties:
        tag_id:
          type: string
          title: Tag Id
        entity_id:
          type: string
          title: Entity Id
        entity_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Entity Name
        entity_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Entity Type
        property:
          type: string
          title: Property
        description:
          type: string
          title: Description
        unit:
          type: string
          title: Unit
        category:
          type: string
          title: Category
      type: object
      required:
      - tag_id
      - entity_id
      - property
      - description
      - unit
      - category
      title: OntologyTagInfo
    RangeResponse:
      properties:
        count:
          type: integer
          title: Count
        truncated:
          type: boolean
          title: Truncated
        snapshots:
          items:
            $ref: '#/components/schemas/Snapshot'
          type: array
          title: Snapshots
      type: object
      required:
      - count
      - truncated
      - snapshots
      title: RangeResponse
    Snapshot:
      properties:
        timestamp:
          type: string
          title: Timestamp
        cycle:
          type: integer
          title: Cycle
        solved:
          type: boolean
          title: Solved
        solve_time_s:
          anyOf:
          - type: number
          - type: 'null'
          title: Solve Time S
        cycle_duration_s:
          anyOf:
          - type: number
          - type: 'null'
          title: Cycle Duration S
        tag_count:
          type: integer
          title: Tag Count
          default: 0
        errors:
          items:
            type: string
          type: array
          title: Errors
        tags:
          additionalProperties: true
          type: object
          title: Tags
        errors_truncated:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Errors Truncated
        total_error_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Total Error Count
      type: object
      required:
      - timestamp
      - cycle
      - solved
      title: Snapshot
    TagEntry:
      properties:
        tag_id:
          type: string
          title: Tag Id
        owner_tag:
          type: string
          title: Owner Tag
        owner_type:
          type: string
          title: Owner Type
        phase:
          anyOf:
          - type: string
          - type: 'null'
          title: Phase
        property_key:
          type: string
          title: Property Key
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        unit_si:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit Si
        category:
          type: string
          title: Category
        subsystem:
          anyOf:
          - type: string
          - type: 'null'
          title: Subsystem
        property_package:
          anyOf:
          - type: string
          - type: 'null'
          title: Property Package
        static_composition:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Static Composition
        composition_meaningful:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Composition Meaningful
        current_value:
          title: Current Value
      type: object
      required:
      - tag_id
      - owner_tag
      - owner_type
      - property_key
      - category
      title: TagEntry
    TagHistoryPoint:
      properties:
        timestamp:
          type: string
          title: Timestamp
        value:
          title: Value
      type: object
      required:
      - timestamp
      title: TagHistoryPoint
    TagHistoryResponse:
      properties:
        tag_id:
          type: string
          title: Tag Id
        count:
          type: integer
          title: Count
        points:
          items:
            $ref: '#/components/schemas/TagHistoryPoint'
          type: array
          title: Points
      type: object
      required:
      - tag_id
      - count
      - points
      title: TagHistoryResponse
    TagValueResponse:
      properties:
        tag_id:
          type: string
          title: Tag Id
        timestamp:
          type: string
          title: Timestamp
        cycle:
          type: integer
          title: Cycle
        value:
          title: Value
      type: object
      required:
      - tag_id
      - timestamp
      - cycle
      title: TagValueResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError